pyproject.toml
[project]
name = "patanode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.5"
dependencies = [
    "bytechomp",
    "gif2numpy>=1.3",
    "ifaddr>=0.2.0",
    "keyboard>=0.13.5",
    "librosa>=0.11.0",
    "libusb-package>=1.0.26.3",
    "matplotlib>=3.10.3",
    "moderngl>=5.12.0",
    "nodeeditor",
    "opencv-python-headless>=4.11.0.86",
    "pydantic>=2.11.7",
    "pygame>=2.6.1",
    "pyglm>=2.8.2",
    "pygltflib>=1.16.4",
    "pyqt5>=5.15.11",
    "pyqtgraph>=0.13.7",
    "pyrr>=0.10.3",
    "pyusb>=1.3.1",
    "pyyaml>=6.0.2",
    "scipy>=1.16.0",
    "sounddevice>=0.5.2",
]

[tool.uv.sources]
nodeeditor = { git = "https://github.com/pata-tech/pyqt-node-editor", rev = "master" }
bytechomp = { git = "https://github.com/theof/bytechomp.git", rev = "main" }

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]

[tool.uv]
constraint-dependencies = ["pyqt5-qt5 <=5.15.2"]
